Adjusting configuration and adding documentation
