--- example/src/main/resources/application.properties
@@ -1,9 +1,0 @@
-spring.application.name=example
-
-
-#spring.cloud.kubernetes.reload.strategy=refresh
-#spring.cloud.kubernetes.reload.mode=polling
-#spring.cloud.kubernetes.reload.period=2000
-#spring.cloud.kubernetes.reload.monitoring-config-maps=false
-#spring.cloud.kubernetes.reload.monitoring-secrets=true
-#spring.cloud.kubernetes.reload.enabled=true
