+++ spring-cloud-kubernetes-examples/kubernetes-reload-example/config-map.yml
@@ -1,0 +1,8 @@
+apiVersion: v1
+kind: ConfigMap
+metadata:
+  name: reload-example
+data:
+  application.properties: |-
+    bean.message=Hello World!
+    another.property=value
