+++ spring-cloud-kubernetes-examples/kubernetes-reload-example/src/main/java/io/fabric8/spring/cloud/kubernetes/examples/App.java
@@ -1,12 +1,14 @@
-package io.app;
+package io.fabric8.spring.cloud.kubernetes.examples;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.scheduling.annotation.EnableScheduling;
 
 /**
  *
  */
 @SpringBootApplication
+@EnableScheduling
 public class App {
 
     public static void main(String[] args) {
