+++ spring-cloud-kubernetes-examples/kubernetes-reload-example/src/main/resources/application.properties
@@ -1,0 +1,9 @@
+spring.application.name=reload-example
+
+spring.cloud.kubernetes.reload.enabled=true
+
+#spring.cloud.kubernetes.reload.strategy=restart_context
+#spring.cloud.kubernetes.reload.mode=polling
+#spring.cloud.kubernetes.reload.period=5000
+#spring.cloud.kubernetes.reload.monitoring-config-maps=false
+#spring.cloud.kubernetes.reload.monitoring-secrets=true
