+++ spring-cloud-kubernetes-examples/pom.xml
@@ -1,0 +1,23 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <parent>
+        <artifactId>spring-cloud-kubernetes-project</artifactId>
+        <groupId>io.fabric8</groupId>
+        <version>0.1-SNAPSHOT</version>
+    </parent>
+    <modelVersion>4.0.0</modelVersion>
+
+    <artifactId>spring-cloud-kubernetes-examples</artifactId>
+    <packaging>pom</packaging>
+
+    <name>Fabric8 :: Spring Cloud Kubernetes :: Examples</name>
+    <description>Examples showing how to use features of Spring Cloud Kubernetes.</description>
+
+    <modules>
+        <module>kubernetes-reload-example</module>
+    </modules>
+
+
+</project>
